Renamed file and module liststyleui -> listpropertiesui
